+++ R/matchSpectra.R
@@ -15,8 +15,9 @@
 #' the similarity calculation and matching condition. Parameter `query` takes
 #' a [Spectra] object while `target` can be either a [Spectra] object, a
 #' [CompDb] (reference library) object defined in the `CompoundDb` package or
-#' a [CompAnnotationSource] (e.g. a [CompDbSource()] with the reference or
-#' connection information to a supported annotation resource).
+#' a [CompAnnotationSource] (e.g. a [CompDbSource()] or a [SpectraDbSource()]
+#' with the reference or connection information to a supported annotation
+#' resource).
 #'
 #' Currently supported parameter objects defining the matching are:
 #'
@@ -114,8 +115,8 @@
 #'   for query spectra with missing precursor m/z.
 #'
 #' @param target for `matchSpectra`: [Spectra], [CompDb] or object extending
-#'   [CompAnnotationSource] (such as [CompDbSource]) with the target
-#'   (reference) spectra to compare `query` against.
+#'   [CompAnnotationSource] (such as [CompDbSource] or [SpectraDbSource]) with
+#'   the target (reference) spectra to compare `query` against.
 #'
 #' @param tolerance `numeric(1)` for an absolute maximal accepted difference
 #'   between m/z values. This will be used in `compareSpectra` as well as for